implement no grad to optimize the cob and reduce pred_difference

+++ neuralteleportation/changeofbasisutils.py
@@ -37,12 +37,12 @@ def get_random_cob(range_cob: float, size: int, sampling_type: str = 'intra_land
         if center != 1:
             print('Warning: The change of basis sampling is not centered at 1. But no negative change of basis'
                   ' will be produced.')
-        cob = np.random.uniform(low=-range_cob, high=range_cob, size=size).astype(np.float) + 1
+        cob = np.random.uniform(low=-range_cob, high=range_cob, size=size).astype(np.float32) + 1
 
     # Change of basis equally in intervals [-1-range_cob, -1+range_cob] and [1-range_cob, 1+range_cob]
     elif sampling_type == 'inter_landscape':
         samples = np.random.randint(0, 2, size=size)
-        cob = np.zeros_like(samples, dtype=np.float)
+        cob = np.zeros_like(samples, dtype=np.float32)
         cob[samples == 1] = np.random.uniform(low=-1-range_cob, high=-1+range_cob, size=samples.sum())
         cob[samples == 0] = np.random.uniform(low=1-range_cob, high=1+range_cob, size=(len(samples) - samples.sum()))
 
@@ -50,15 +50,15 @@ def get_random_cob(range_cob: float, size: int, sampling_type: str = 'intra_land
     elif sampling_type == 'centered':
         assert not (range_cob > center or center <= 0), 'This range for change of basis sampling allows for negative ' \
                                                         'changes of basis.'
-        cob = np.random.uniform(low=center - range_cob, high=center + range_cob, size=size).astype(np.float)
+        cob = np.random.uniform(low=center - range_cob, high=center + range_cob, size=size).astype(np.float32)
 
     # Change of basis in interval [0, range_cob]
     elif sampling_type == 'positive':
-        cob = np.random.uniform(low=0, high=range_cob, size=size).astype(np.float)
+        cob = np.random.uniform(low=0, high=range_cob, size=size).astype(np.float32)
 
     # Change of basis in interval [-range_cob, 0]
     elif sampling_type == 'negative':
-        cob = np.random.uniform(low=-range_cob, high=0, size=size).astype(np.float)
+        cob = np.random.uniform(low=-range_cob, high=0, size=size).astype(np.float32)
 
     else:
         raise ValueError("Invalid sampling type. Sampling types allowed: "
